+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+
+0.2.1 (2023-11-??)
+-------------------------
+* Update Manney et al. 2011 method to include 'check_diagonals' parameter
+
+
 0.2.0 (2023-10-12)
 -------------------------
 * Fix CI and Github actions